+++ config/model-reflection.php
@@ -1,6 +1,0 @@
-<?php
-
-// config for Spatie/ModelReflection
-return [
-
-];