+++ tests/ModelMetaTest.php
@@ -1,8 +1,8 @@
 <?php
 
-use Spatie\ModelReflection\ModelMeta;
-use Spatie\ModelReflection\Tests\TestSupport\Models\RelationTestModel;
-use Spatie\ModelReflection\Tests\TestSupport\Models\TestModel;
+use Spatie\ModelMeta\ModelMeta;
+use Spatie\ModelMeta\Tests\TestSupport\Models\RelationTestModel;
+use Spatie\ModelMeta\Tests\TestSupport\Models\TestModel;
 
 it('can get meta information about a model', function() {
     $modelMeta = ModelMeta::forModel(RelationTestModel::class);