+++ config/model-meta.php
@@ -1,6 +1,0 @@
-<?php
-
-// config for Spatie/ModelMeta
-return [
-
-];