+++ tests/TestSupport/TestCase.php
@@ -1,33 +1,17 @@
 <?php
 
-namespace Spatie\ModelMeta\Tests\TestSupport;
+namespace Spatie\ModelInfo\Tests\TestSupport;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Orchestra\Testbench\TestCase as Orchestra;
-use Spatie\ModelMeta\ModelMetaServiceProvider;
+use Spatie\ModelInfo\ModelInfoServiceProvider;
 use Spatie\Snapshots\MatchesSnapshots;
 
 class TestCase extends Orchestra
 {
     use MatchesSnapshots;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'Spatie\\ModelMeta\\Database\\Factories\\'.class_basename($modelName).'Factory'
-        );
-    }
-
-    protected function getPackageProviders($app)
-    {
-        return [
-            ModelMetaServiceProvider::class,
-        ];
-    }
 
     public function getEnvironmentSetUp($app)
     {