+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 All notable changes to `laravel-model-info` will be documented in this file.
+
+## 0.0.2 - 2022-08-09
+
+**Full Changelog**: https://github.com/spatie/laravel-model-info/compare/0.0.1...0.0.2